+++ src/main/java/nextstep/ladder/domain/LadderWidth.java
@@ -2,6 +2,7 @@ package nextstep.ladder.domain;
 
 import java.util.ArrayList;
 import java.util.List;
+import nextstep.ladder.interfaces.Position;
 
 public class LadderWidth {
     private static final MinSizeValidator minSizeValidator = new MinSizeValidator();
