+++ src/main/java/nextstep/ladder/domain/LineImpl.java
@@ -1,40 +1,33 @@
 package nextstep.ladder.domain;
 
 import java.util.Objects;
+import nextstep.ladder.interfaces.Line;
+import nextstep.ladder.interfaces.Position;
 
-public class Line {
+public class LineImpl implements Line {
 
     private final LineRange lineRange;
     private boolean hasPoint;
 
-    public Line(boolean hasPoint, int left) {
+    public LineImpl(boolean hasPoint, int left) {
         this(hasPoint, new LineRange(left));
     }
 
-    public Line(boolean hasPoint, LineRange lineRange) {
+    public LineImpl(boolean hasPoint, LineRange lineRange) {
         this.hasPoint = hasPoint;
         this.lineRange = lineRange;
     }
 
+    @Override
     public void move(Position position) {
         if (hasPoint) {
             lineRange.swap(position);
         }
     }
 
+    @Override
     public boolean canMove() {
         return hasPoint;
-    }
-
-    public boolean connected(Line line) {
-        if (hasNoConsecutivePoints(line)) {
-            return false;
-        }
-        return lineRange.isContinue(line.lineRange);
-    }
-
-    private boolean hasNoConsecutivePoints(Line line) {
-        return !hasPoint || !line.hasPoint;
     }
 
     @Override
@@ -45,7 +38,7 @@ public class Line {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        Line point = (Line) o;
+        LineImpl point = (LineImpl) o;
         return hasPoint == point.hasPoint && Objects.equals(lineRange, point.lineRange);
     }
