+++ src/main/java/nextstep/ladder/factory/LinesFactory.java
@@ -1,0 +1,20 @@
+package nextstep.ladder.factory;
+
+import java.util.Arrays;
+import java.util.List;
+import nextstep.ladder.domain.Lines;
+
+public class LinesFactory {
+    public static Lines lines(List<Boolean> lines) {
+        return new Lines(LinesDataConverter.convert(lines));
+    }
+
+    public static Lines lines(Boolean... lines) {
+        return lines(toList(lines));
+    }
+
+    private static List<Boolean> toList(Boolean[] lines) {
+        return Arrays.asList(lines);
+    }
+
+}
