+++ src/main/java/nextstep/ladder/interfaces/Line.java
@@ -1,0 +1,7 @@
+package nextstep.ladder.interfaces;
+
+public interface Line {
+    void move(Position position);
+
+    boolean canMove();
+}
